Pre-processing steps
The following code describes the overall filtration process for the scRNA-seq analysis. Libraries loaded are for the entire analysis. The first step involves QC filtration by identifying the sum of cell counts across the matrix. This is followed by creating the Seurat object where all genes which are expressed in less than 5 cells and all cells which express less than 200 features are removed. Data is visualized for number of genes, cell counts and mitochondrial DNA content. Final cell-level filtration is set as:
1) Minimum features 250 and maximum features at 95th percentile.
2) Minimum counts at 500 and maximum counts at 95th percentile
3) Mitochondrial DNA percent of not more than 10%
4) log UMI (Unique Molecular identified) value of greater than 0.8