Normalization and Finding variable features
Object from pre-processing is subjected to Log Normalization. Top 2000 variable features are identified using the vst method. The expression of top 10 genes are visualized. Genes of interest are chosen based in the TGFB downstream pathway and genes falling within highly variable are identified. Data is then scaled for all genes prior to dimensionality reduction and expression of housekeeping genes GAPDH and ACTB are checked in the metadata